your-odyssey-backend/trip_planner/requirements.txt
Flask==2.3.3
pydantic-ai==0.0.14
google-generativeai==0.8.3
google-api-python-client==2.108.0
requests==2.31.0
python-dotenv==1.0.0